SleepScore/StateEditor update

+++ preprocessing/bz_PreprocessExtracellEphysSession.m
@@ -12,15 +12,19 @@ bz_RunSessionMetadata(basepath);
 %% making an xml if needed
 basename = bz_BasenameFromBasepath(basepath);
 txmlname = fullfile(basepath,[basename '.xml']);
-d = dir(fullfile(basepath,txmlname));
+d = dir(txmlname);
 load(fullfile(basepath,[basename '.SessionMetadata.mat']))
 if isempty(d);%if no basename.xml already
     if isempty(SessionMetadata.ExtracellEphys.ParametersDivergentFromAnimalMetadata)%if not different from animal xml, copy the animal xml here
         ax = fullfile(SessionMetadata.AnimalMetadata.AnimalBasepath,[SessionMetadata.AnimalMetadata.AnimalName, '.xml']);
-        copyfile(ax,txmlname);
-    else    %if changes, make new xml for this session
-        bz_MakeXML(basepath)
+        if exist(ax,'file')
+            copyfile(ax,txmlname);
+        end
     end
+end
+d = dir(txmlname);
+if isempty(d);%if no basename.xml already
+    bz_MakeXML(basepath)
 end
 
 %%
